fixed dir and component name rule.
The directory structure under src follows the directory structure of `@madogiwa-ui/css` to make the target elements easy to understand. For example, the component for `.m-block` in `components/block-name` directory -> `MBlcok` component in `components/MBlock` directory.

However, in cases where the block name and the file name do not match (for example, when `button.css` targets `.m-btn`), use the CSS file name as the directory name and use the block name for the component name. For example: `button.css` (for `.m-btn`) -> `MButton/MBtn.vue`.